+++ src/combinator/decorator/value.rs
@@ -5,49 +5,14 @@ use crate::{
   combinator::{Action, Combinator, Input, Output},
   range::WithRange,
 };
-use core::fmt;
-use std::marker::PhantomData;
 
 create_closure_decorator!(Map, "See [`Combinator::map`].");
 create_simple_decorator!(Tuple, "See [`Combinator::tuple`].");
 create_value_decorator!(Bind, "See [`Combinator::bind`].");
+create_closure_decorator!(BindWith, "See [`Combinator::bind_with`].");
 create_closure_decorator!(Select, "See [`Combinator::select`].");
 create_simple_decorator!(Range, "See [`Combinator::range`].");
 create_simple_decorator!(Pop, "See [`Combinator::pop`].");
-
-/// See [`Combinator::bind_default`].
-pub struct BindDefault<T, D> {
-  action: T,
-  inner: PhantomData<D>,
-}
-
-impl<T: Copy, D> Copy for BindDefault<T, D> {}
-impl<T: Clone, D> Clone for BindDefault<T, D> {
-  fn clone(&self) -> Self {
-    Self {
-      action: self.action.clone(),
-      inner: PhantomData,
-    }
-  }
-}
-impl<T: fmt::Debug, D> fmt::Debug for BindDefault<T, D> {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    let action = &self.action;
-    f.debug_struct(stringify!(BindDefault))
-      .field(stringify!(action), action)
-      .finish()
-  }
-}
-
-impl<T, D> BindDefault<T, D> {
-  #[inline]
-  const fn new(action: T) -> Self {
-    Self {
-      action,
-      inner: PhantomData,
-    }
-  }
-}
 
 unsafe impl<
     Text: ?Sized,
@@ -94,17 +59,17 @@ unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap>, D: Clone>
   }
 }
 
-unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap>, D: Default>
-  Action<Text, State, Heap> for BindDefault<T, D>
+unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap>, NewValue, D: Fn() -> NewValue>
+  Action<Text, State, Heap> for BindWith<T, D>
 {
-  type Value = D;
+  type Value = NewValue;
 
   #[inline]
   fn exec(&self, input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self
       .action
       .exec(input)
-      .map(|output| output.map(|_| Default::default()))
+      .map(|output| output.map(|_| (self.inner)()))
   }
 }
 
@@ -229,17 +194,20 @@ impl<T> Combinator<T> {
     Combinator::new(Bind::new(self.action, value))
   }
 
-  /// Create a new combinator to set [`Output::value`] to the default value.
+  /// Create a new combinator to set [`Output::value`] with the provided factory.
   /// # Examples
   /// ```
   /// # use whitehole::{action::Action, combinator::Combinator};
   /// # fn t(combinator: Combinator<impl Action>) -> Combinator<impl Action<Value=i32>> {
-  /// combinator.bind_default()
+  /// combinator.bind_with(|| 0i32)
   /// # }
   /// ```
   #[inline]
-  pub fn bind_default<NewValue: Default>(self) -> Combinator<BindDefault<T, NewValue>> {
-    Combinator::new(BindDefault::new(self.action))
+  pub fn bind_with<NewValue, Factory: Fn() -> NewValue>(
+    self,
+    factory: Factory,
+  ) -> Combinator<BindWith<T, Factory>> {
+    Combinator::new(BindWith::new(self.action, factory))
   }
 
   /// Create a new combinator to set [`Output::value`] by the `selector`.
@@ -386,10 +354,10 @@ mod tests {
   }
 
   #[test]
-  fn combinator_bind_default() {
+  fn combinator_bind_with() {
     assert_eq!(
       wrap(|input| input.digest(1))
-        .bind_default::<i32>()
+        .bind_with(|| 0i32)
         .exec(Input::new(Instant::new("123"), &mut (), &mut ())),
       Some(Output {
         value: 0,
@@ -398,7 +366,7 @@ mod tests {
     );
     assert_eq!(
       bytes::wrap(|input| input.digest(1))
-        .bind_default::<i32>()
+        .bind_with(|| 0i32)
         .exec(Input::new(Instant::new(b"123"), &mut (), &mut ())),
       Some(Output {
         value: 0,
@@ -407,13 +375,13 @@ mod tests {
     );
 
     // make sure copy-able and clone-able
-    let a = wrap::<(), (), _, _>(|input| input.digest(1)).bind_default::<i32>();
+    let a = wrap::<(), (), _, _>(|input| input.digest(1)).bind_with(|| 0i32);
     let _ = a;
     let _ = a.clone();
 
     assert_eq!(
       format!("{:?}", a),
-      "Combinator { action: BindDefault { action: Wrap } }"
+      "Combinator { action: BindWith { action: Wrap } }"
     );
   }