+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `Combinator::bind_with`.
+
+### Removed
+
+- Remove `Combinator::bind_default`, use `Combinator::bind_with` instead.
+
 ## [0.5.0] - 2025-01-28
 
 ### Added